Consolidated code into RuleNetworkEvaluator
