--- drools-core/src/main/java/org/drools/core/phreak/TupleEvaluationUtil.java
@@ -1,57 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.drools.core.phreak;
-
-
-import org.drools.core.common.ReteEvaluator;
-import org.drools.core.reteoo.PathMemory;
-import org.drools.core.reteoo.SegmentMemory;
-import org.drools.core.reteoo.Tuple;
-import org.drools.core.reteoo.TupleImpl;
-
-import static org.drools.core.common.TupleSetsImpl.createLeftTupleTupleSets;
-
-public class TupleEvaluationUtil {
-
-    public static boolean flushLeftTupleIfNecessary(ReteEvaluator reteEvaluator, SegmentMemory sm, boolean streamMode) {
-        return flushLeftTupleIfNecessary(reteEvaluator, sm, null, streamMode, Tuple.NONE);
-    }
-
-    public static boolean flushLeftTupleIfNecessary(ReteEvaluator reteEvaluator,
-                                                    SegmentMemory sm,
-                                                    TupleImpl leftTuple,
-                                                    boolean streamMode,
-                                                    short stagedType) {
-        PathMemory pmem = findPathToFlush(sm, leftTuple, streamMode);
-
-        if (pmem == null) {
-            return false;
-        }
-
-        reteEvaluator.getRuleNetworkEvaluator().forceFlushLeftTuple(pmem, sm, createLeftTupleTupleSets(leftTuple, stagedType));
-        reteEvaluator.getRuleNetworkEvaluator().forceFlushWhenSubnetwork(pmem);
-        return true;
-    }
-
-    public static PathMemory findPathToFlush(SegmentMemory sm, TupleImpl leftTuple, boolean streamMode) {
-        boolean forceFlush = streamMode || (leftTuple != null && leftTuple.getFactHandle() != null && leftTuple
-                .getFactHandle().isEvent());
-        return forceFlush ? sm.getPathMemories().get(0) : sm.getFirstDataDrivenPathMemory();
-    }
-}
