+++ drools-core/src/main/java/org/drools/core/reteoo/LeftInputAdapterNode.java
@@ -43,6 +43,7 @@ import org.drools.core.common.TupleSets;
 import org.drools.core.common.TupleSetsImpl;
 import org.drools.core.common.UpdateContext;
 import org.drools.core.phreak.DetachedTuple;
+import org.drools.core.phreak.RuleNetworkEvaluatorImpl;
 import org.drools.core.reteoo.builder.BuildContext;
 import org.drools.core.rule.consequence.InternalMatch;
 import org.drools.core.util.AbstractLinkedListNode;
@@ -53,8 +54,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.drools.base.reteoo.PropertySpecificUtil.isPropertyReactive;
-import static org.drools.core.phreak.TupleEvaluationUtil.findPathToFlush;
-import static org.drools.core.phreak.TupleEvaluationUtil.flushLeftTupleIfNecessary;
 
 /**
  * All asserting Facts must propagated into the right <code>ObjectSink</code> side of a BetaNode, if this is the first Pattern
@@ -237,7 +236,7 @@ public class LeftInputAdapterNode extends LeftTupleSource
     }
 
     public static List<PathMemory> doInsertSegmentMemory(ReteEvaluator reteEvaluator, boolean linkOrNotify, LiaNodeMemory lm, SegmentMemory sm, TupleImpl leftTuple, boolean streamMode) {
-        PathMemory pmem = findPathToFlush(sm, leftTuple, streamMode);
+        PathMemory pmem = RuleNetworkEvaluatorImpl.findPathToFlush(sm, leftTuple, streamMode);
         if ( pmem != null ) {
             reteEvaluator.getRuleNetworkEvaluator().forceFlushLeftTuple(pmem, sm, TupleSetsImpl.createLeftTupleTupleSets(leftTuple, Tuple.INSERT));
             if ( linkOrNotify ) {
@@ -298,7 +297,7 @@ public class LeftInputAdapterNode extends LeftTupleSource
     private static void doDeleteSegmentMemory(TupleImpl leftTuple, PropagationContext pctx, final LiaNodeMemory lm,
                                               SegmentMemory sm, ReteEvaluator reteEvaluator, boolean linkOrNotify, boolean streamMode) {
         leftTuple.setPropagationContext( pctx );
-        if ( flushLeftTupleIfNecessary( reteEvaluator, sm, leftTuple, streamMode, Tuple.DELETE ) ) {
+        if ( reteEvaluator.getRuleNetworkEvaluator().flushLeftTupleIfNecessary(sm, leftTuple, streamMode, Tuple.DELETE) ) {
             if ( linkOrNotify ) {
                 lm.setNodeDirty( );
             }
@@ -348,7 +347,7 @@ public class LeftInputAdapterNode extends LeftTupleSource
         TupleSets leftTuples = sm.getStagedLeftTuples();
 
         if ( leftTuple.getStagedType() == LeftTuple.NONE ) {
-            if ( flushLeftTupleIfNecessary( reteEvaluator, sm, leftTuple, streamMode, Tuple.UPDATE ) ) {
+            if ( reteEvaluator.getRuleNetworkEvaluator().flushLeftTupleIfNecessary(sm, leftTuple, streamMode, Tuple.UPDATE) ) {
                 if ( linkOrNotify ) {
                     lm.setNodeDirty( );
                 }
